+++ postcss.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  plugins: {
+    'postcss-cssnext': {},
+    'postcss-url': {
+      url: 'inline',
+    },
+    stylefmt: {},
+  },
+};