+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
 
+## 0.7.2 (Aug 19, 2017)
+
+- Ignore mouse down event when the viewer is hiding (#70).
+- Fixed multiple active items in navbar (#75).
+
+
+## 0.7.1 (May 14, 2017)
+
+- Support to use Viewer in a modal (#39).
+
+
 ## 0.7.0 (Apr 30, 2017)
 
 - Changed the `main` field value from `dist/viewer.js` (UMD) to `dist/viewer.common.js` (CommonJS).